image_recognition/requirements.txt
opencv-python==4.8.1
numpy==1.24.3
Pillow==10.0.0
tensorflow==2.13.0
paddlepaddle==2.5.1
paddleocr==2.7.0.3
python-docx==0.8.11 